+++ resources/views/admin/user/update.php
@@ -9,7 +9,7 @@ declare(strict_types=1);
  * @var UserForm $form
  */
 
-use Hirtz\Skeleton\Modules\Admin\Models\forms\UserForm;
+use Hirtz\Skeleton\Modules\Admin\Models\Forms\UserForm;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Forms\UserActiveForm;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Navs\UserHeader;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Navs\UserSubmenu;